+++ src/test/utility_test.rs
@@ -8,5 +8,6 @@ mod label_encoding_test;
 mod linear_discriminant_analysis_test;
 mod normalize_test;
 mod principal_component_analysis_test;
+mod standardize_test;
 mod t_sne_test;
 mod train_test_split_test;