+++ changelog.md
@@ -13,6 +13,7 @@ Please view [SomeB1oody/RustyML](https://github.com/SomeB1oody/RustyML) for more
 - Refactor `param_count` method across layers to use `TrainingParameters` enum
 - Refactor dataset loading functions to return merged headers and rows as single `&'static str`
 - Update `normalize` function to accept `ArrayBase`
+- Implement `standardize` function with flexible axis-based standardization (Row/Column/Global)
 
 ## [v0.7.0] - 2025-9-26 (UTC-7)
 ### Changed